+++ ProSoft/EasySave/src/Models/Data/SaveType.cs
@@ -1,8 +1,18 @@
 ﻿namespace EasySave.src.Models.Data
 {
+    /// <summary>
+    /// Save type enum
+    /// </summary>
     public enum SaveType
     {
+        /// <summary>
+        /// Full save
+        /// </summary>
         Full,
+
+        /// <summary>
+        /// Differential save
+        /// </summary>
         Differential,
     }
 }
