--- ProSoft/EasySave/src/Models/Dir.cs
@@ -1,8 +1,0 @@
-﻿namespace EasySave.src.Models
-{
-    public interface IDir
-    {
-        string Path { get; }
-    }
-
-}
