+++ ProSoft/EasySave/src/Models/Exceptions/CantCheckUpdateException.cs
@@ -3,9 +3,15 @@
     using System;
     using System.Runtime.Serialization;
 
+    /// <summary>
+    /// Exception thrown when the update can't be checked
+    /// </summary>
     [Serializable]
     public class CantCheckUpdateException : Exception
     {
+
+        //All constructors are required by [Serializable]
+
         public CantCheckUpdateException()
         {
         }
