+++ ProSoft/EasySave/src/Models/IDir.cs
@@ -1,0 +1,15 @@
+﻿namespace EasySave.src.Models
+{
+    /// <summary>
+    /// Interface for directory
+    /// </summary>
+    public interface IDir
+    {
+
+        /// <summary>
+        /// Path of the directory
+        /// </summary>
+        string Path { get; }
+    }
+
+}
