+++ ProSoft/EasySave/src/Models/SrcDir.cs
@@ -1,23 +1,30 @@
-﻿using EasySave.Properties;
-using EasySave.src.Utils;
-using Spectre.Console;
-using System;
-using System.Collections.Generic;
+﻿using EasySave.src.Utils;
 using System.IO;
-using System.Text;
-using static System.Net.WebRequestMethods;
 
 namespace EasySave.src.Models
 {
+    /// <summary>
+    /// Source directory, implements IDir
+    /// </summary>
     public class SrcDir : IDir
     {
 
         public string Path { get; }
 
+        /// <summary>
+        /// Size of the directory
+        /// </summary>
         private readonly double _size;
 
+        /// <summary>
+        /// Number of files in the directory
+        /// </summary>
         public readonly long NbFiles;
 
+        /// <summary>
+        /// Constructor
+        /// </summary>
+        /// <param name="path">path of dir</param>
         public SrcDir(string path)
         {
             Path = path;
@@ -26,6 +33,10 @@ namespace EasySave.src.Models
             NbFiles = DirectoryUtils.GetDirectoryFiles(directory);
         }
 
+        /// <summary>
+        /// Get the size of the directory
+        /// </summary>
+        /// <returns>size of the directory</returns>
         public double GetSize()
         {
             return _size;
