+++ ProSoft/EasySave/src/Render/RenderMethod.cs
@@ -1,12 +1,38 @@
 ﻿namespace EasySave.src.Render
 {
+    /// <summary>
+    /// Render method enum
+    /// </summary>
     public enum RenderMethod
     {
+        /// <summary>
+        /// Render home
+        /// </summary>
         Home,
+
+        /// <summary>
+        /// Render create save
+        /// </summary>
         CreateSave,
+
+        /// <summary>
+        /// Render load save
+        /// </summary>
         LoadSave,
+
+        /// <summary>
+        /// Render edit save
+        /// </summary>
         EditSave,
+
+        /// <summary>
+        /// Render delete save
+        /// </summary>
         DeleteSave,
+
+        /// <summary>
+        /// Render change language
+        /// </summary>
         ChangeLanguage
     }
 }
