+++ ProSoft/EasySave/src/Utils/DirectoryUtils.cs
@@ -6,32 +6,53 @@ using System.Threading.Tasks;
 
 namespace EasySave.src.Utils
 {
+    /// <summary>
+    /// Static class to manage directory actions
+    /// </summary>
     public static class DirectoryUtils
     {
 
+        /// <summary>
+        /// Array to store the actual file being copied
+        /// </summary>
         private static readonly string[] actualFile = new string[2];
 
-        public static bool CopyFilesAndFolders(Save s, SaveType type = SaveType.Full)
+        /// <summary>
+        /// Copy all files and folders from a source directory to a destination directory
+        /// </summary>
+        /// <param name="s">save concerned</param>
+        /// <returns></returns>
+        public static void CopyFilesAndFolders(Save s)
         {
             DirectoryInfo sourceDirectory = new DirectoryInfo(s.SrcDir.Path);
             DirectoryInfo destinationDirectory = new DirectoryInfo(s.DestDir.Path);
+            //Parallel is used to display progress bar while data beeing copied
             Parallel.Invoke(
                 () => ConsoleUtils.CreateProgressBar(s),
-                () => CopyAll(s, sourceDirectory, destinationDirectory, type)
+                () => CopyAll(s, sourceDirectory, destinationDirectory, s.GetSaveType())
             );
-            return true;
         }
 
+        /// <summary>
+        /// Method to copy all files and folders from a source directory to a destination directory
+        /// </summary>
+        /// <param name="s">concerned save</param>
+        /// <param name="src">source dir</param>
+        /// <param name="dest">destination dir</param>
+        /// <param name="type">type of save</param>
         private static void CopyAll(Save s, DirectoryInfo src, DirectoryInfo dest, SaveType type)
         {
             foreach (FileInfo file in src.GetFiles())
             {
+                //Update json data
                 LogUtils.LogSaves();
                 bool fileExists = File.Exists(Path.Combine(dest.FullName, file.Name));
+                //Proceed differential mode by comparing files data
                 if (type == SaveType.Full || !fileExists || (DateTime.Compare(File.GetLastWriteTime(Path.Combine(dest.FullName, file.Name)), File.GetLastWriteTime(Path.Combine(src.FullName, file.Name))) < 0))
                 {
                     actualFile[0] = src.FullName;
                     actualFile[1] = dest.FullName;
+                    //Stopwatch to mesure transfer time
                     var watch = new System.Diagnostics.Stopwatch();
                     watch.Start();
                     try
@@ -43,12 +64,14 @@ namespace EasySave.src.Utils
                         ConsoleUtils.WriteError($"{file.Name} | {Resource.AccesDenied}");
                     }
                     watch.Stop();
+                    //Log transfer in json
                     LogUtils.LogTransfer(s, Path.Combine(src.FullName, file.Name), Path.Combine(dest.FullName, file.Name), file.Length, watch.ElapsedMilliseconds);
                 }
                 s.AddFileCopied();
                 s.AddSizeCopied(file.Length);
             }
 
+            //Recursive call for subdirectories
             foreach (DirectoryInfo directory in src.GetDirectories())
             {
                 DirectoryInfo nextTarget = dest.CreateSubdirectory(directory.Name);
@@ -56,16 +79,30 @@ namespace EasySave.src.Utils
             }
         }
 
-        public static bool IsValidPath(String path)
+        /// <summary>
+        /// check if path is valid
+        /// </summary>
+        /// <param name="path">path to test</param>
+        /// <returns>true if valid path, else otherwise</returns>
+        public static bool IsValidPath(string path)
         {
             return Directory.Exists(path);
         }
 
+        /// <summary>
+        /// Create a directory
+        /// </summary>
+        /// <param name="path">path of dir</param>
         public static void CreatePath(string path)
         {
             Directory.CreateDirectory(path);
         }
 
+        /// <summary>
+        /// get directory size method
+        /// </summary>
+        /// <param name="path">directory</param>
+        /// <returns>size of the directory</returns>
         public static double GetDirectorySize(DirectoryInfo path)
         {
             double size = 0;
@@ -76,6 +113,11 @@ namespace EasySave.src.Utils
             return size;
         }
 
+        /// <summary>
+        /// get number of files in a directory
+        /// </summary>
+        /// <param name="path">directory</param>
+        /// <returns>files in directory</returns>
         public static long GetDirectoryFiles(DirectoryInfo path)
         {
             long nbFiles = 0;
@@ -86,6 +128,10 @@ namespace EasySave.src.Utils
             return nbFiles;
         }
 
+        /// <summary>
+        /// get actual file being copied
+        /// </summary>
+        /// <returns>actual file</returns>
         public static string[] GetActualFile()
         {
             return actualFile;
