+++ ProSoft/EasySave/src/Utils/LogUtils.cs
@@ -7,20 +7,34 @@ using System.IO;
 
 namespace EasySave.src.Utils
 {
+    /// <summary>
+    /// Log utils class
+    /// </summary>
     public static class LogUtils
     {
 
+        /// <summary>
+        /// Path to the log file
+        /// </summary>
         private static readonly string _path = $@"{Environment.GetFolderPath(Environment.SpecialFolder.ApplicationData)}\EasySave\";
 
+        /// <summary>
+        /// Date of the day
+        /// </summary>
         private static readonly string _date = DateTime.Now.ToString("yyyyMMdd");
 
+        /// <summary>
+        /// Init logs util
+        /// </summary>
         public static void Init()
         {
+            //Create log file if not exists
             if (!DirectoryUtils.IsValidPath(_path))
             {
                 AnsiConsole.Clear();
                 Directory.CreateDirectory(_path);
             }
+            //If file exists, load saves
             if (File.Exists($"{_path}saves.json"))
             {
                 try
@@ -35,11 +49,18 @@ namespace EasySave.src.Utils
             }
         }
 
+        /// <summary>
+        /// Method to log saves to json
+        /// </summary>
         public static void LogSaves()
         {
             File.WriteAllText($"{_path}saves.json", SavesToJson().ToString());
         }
 
+        /// <summary>
+        /// Convert saves into json
+        /// </summary>
+        /// <returns>json object</returns>
         private static JObject SavesToJson()
         {
             JObject data = new JObject();
@@ -51,7 +72,12 @@ namespace EasySave.src.Utils
             return data;
         }
 
-        public static dynamic SaveToJson(Save s)
+        /// <summary>
+        /// Convert save into json
+        /// </summary>
+        /// <param name="s">save</param>
+        /// <returns>json object</returns>
+        public static JObject SaveToJson(Save s)
         {
             JobStatus status = s.GetStatus();
             dynamic data = new JObject();
@@ -74,14 +100,22 @@ namespace EasySave.src.Utils
             return data;
         }
 
-        public static void LogTransfer(Save s, string SourcePath, string DestinationPath, long FileSize, float FileTransferTime)
+        /// <summary>
+        /// Log transfer
+        /// </summary>
+        /// <param name="s">save</param>
+        /// <param name="sourcePath">source path</param>
+        /// <param name="destinationPath">destination path</param>
+        /// <param name="fileSize">file size</param>
+        /// <param name="fileTransferTime">file transfer time</param>
+        public static void LogTransfer(Save s, string sourcePath, string destinationPath, long fileSize, float fileTransferTime)
         {
             dynamic transferInfo = new JObject();
             transferInfo.name = $"{s.GetName()} ({s.uuid})";
-            transferInfo.fileSource = SourcePath;
-            transferInfo.fileTarget = DestinationPath;
-            transferInfo.fileSize = FileSize;
-            transferInfo.transferTime = FileTransferTime;
+            transferInfo.fileSource = sourcePath;
+            transferInfo.fileTarget = destinationPath;
+            transferInfo.fileSize = fileSize;
+            transferInfo.transferTime = fileTransferTime;
             transferInfo.date = DateTime.Now;
 
             string json = JsonConvert.SerializeObject(transferInfo);
