+++ ProSoft/EasySave/src/Utils/VersionUtils.cs
@@ -5,9 +5,17 @@ using System.Reflection;
 
 namespace EasySave.src.Utils
 {
+    /// <summary>
+    /// Version utils class
+    /// </summary>
     public static class VersionUtils
     {
 
+        /// <summary>
+        /// Get the latest version from github
+        /// </summary>
+        /// <returns>last version</returns>
+        /// <exception cref="CantCheckUpdateException">throw if random error or no internet</exception>
         public static string GetVersionFromGit()
         {
             try
@@ -21,6 +29,12 @@ namespace EasySave.src.Utils
             }
         }
 
+        /// <summary>
+        /// Get version from string
+        /// e.g. "V1.0.0" => [1, 0, 0]
+        /// </summary>
+        /// <param name="version">string version</param>
+        /// <returns>tab of int for major minor and release</returns>
         public static int[] VersionFromStr(string version)
         {
             if (version.StartsWith("V"))
@@ -29,8 +43,13 @@ namespace EasySave.src.Utils
             return new int[] { int.Parse(versionParts[0]), int.Parse(versionParts[1]), (versionParts.Length > 2 ? int.Parse(versionParts[2]) : 0) };
         }
 
+        /// <summary>
+        /// Compare current version with latest version
+        /// </summary>
+        /// <returns>bool if new version exists</returns>
         public static bool CompareVersions()
         {
+            //Get version from project info
             Version versionData = Assembly.GetExecutingAssembly().GetName().Version;
             int[] current = VersionFromStr($"V{versionData.Major}.{versionData.Minor}.{versionData.Build}");
             int[] latest = VersionFromStr(GetVersionFromGit());
