+++ ProSoft/EasySave/src/ViewModel.cs
@@ -5,42 +5,58 @@ using System.Linq;
 
 namespace EasySave.src
 {
+    /// <summary>
+    /// View model class
+    /// </summary>
     public class ViewModel
     {
 
-        public HashSet<string> GetSaves()
-        {
-            HashSet<string> data = new HashSet<string>();
-            foreach (Save s in Save.GetSaves())
-                data.Add(s.ToString());
-            return data;
-        }
-
+        /// <summary>
+        /// Create save method
+        /// </summary>
+        /// <param name="name">name</param>
+        /// <param name="src">source path</param>
+        /// <param name="dest">destination path</param>
+        /// <param name="type">type of save</param>
+        /// <returns>save object</returns>
         public Save CreateSave(string name, string src, string dest, SaveType type)
         {
             return Save.CreateSave(name, src, dest, type);
         }
 
+        /// <summary>
+        /// Edit save method
+        /// </summary>
+        /// <param name="s">save to edit</param>
+        /// <param name="name">new name</param>
         public void EditSave(Save s, string name)
         {
             s.Rename(name);
         }
 
+        /// <summary>
+        /// Delete save method
+        /// </summary>
+        /// <param name="s">save to deletee</param>
         public void DeleteSave(Save s)
         {
             Save.Delete(s.uuid);
         }
 
-        public HashSet<Save> GetSavesByUuid(HashSet<string> names)
-        {
-            return new HashSet<Save>(Save.GetSaves().Where(save => names.Contains(save.ToString())).ToList());
-        }
+        /// <summary>
+        /// Run save method
+        /// </summary>
+        /// <param name="save">save to run</param>
+        /// <returns>save job result</returns>
 
         public string RunSave(Save save)
         {
             return save.Run();
         }
 
+        /// <summary>
+        /// Stop all saves
+        /// </summary>
         internal void StopAllSaves()
         {
             foreach (Save s in Save.GetSaves())
@@ -50,9 +66,35 @@ namespace EasySave.src
             LogUtils.LogSaves();
         }
 
+        /// <summary>
+        /// check for updates
+        /// </summary>
+        /// <returns>bool if up to date</returns>
         public static bool IsUpToDate()
         {
             return !VersionUtils.CompareVersions();
+        }
+
+        /// <summary>
+        /// Get saves by uuids
+        /// </summary>
+        /// <param name="names">uuids list of saves</param>
+        /// <returns>list of saves</returns>
+        public HashSet<Save> GetSavesByUuid(HashSet<string> names)
+        {
+            return new HashSet<Save>(Save.GetSaves().Where(save => names.Contains(save.ToString())).ToList());
+        }
+
+        /// <summary>
+        /// Get all saves names
+        /// </summary>
+        /// <returns>saves names</returns>
+        public HashSet<string> GetSaves()
+        {
+            HashSet<string> data = new HashSet<string>();
+            foreach (Save s in Save.GetSaves())
+                data.Add(s.ToString());
+            return data;
         }
 
     }
